Use new node-github methods in git-manager
git-manager previously created Promise wrappers around commands in the
node-github API. Now that we've updated the dependency, the library uses
Promises by default and changed some of the names of the API endpoints.

This updates git-manager to work with Promises by default and changes the
appropriate names and fields of API endpoints that have changed.